Add files via upload
I can't reliably debug this because my machine is being strange. I think it has to do with the printMaze function, it's honestly garbage and I'm going to get rid of it. So I was hoping you guys could take a functional look at what I have so far to make sure it's, at the very least, doing what we want it to do. It worked for a few of my test cases but then failed for a few others but again, not sure if this was a visualization problem or logical. Let me know what you guys come up with.